doc/cheatsheet fixes
- Add missing MRFs
Or mentions of MRFs in the main list
whose options were already documented.

- Sort main MRF list alphabetically
It is too cumbersome to find a specific policy atm
or to check if all are docuemtned yet.
Trivial placeholder policies are excluded from this.

- Add missing MRF config detail docs
And remove “on by default” text from individual entries.
They are now laready in the “on by default” section.

- Split out always active MRFs
It doesn’t make sense to add/remove them from the policies list

- Add link to ActivityExpiration config details

- Move MRF policies key to end of section
This makes it easier to spot the transparency options

- Move :activitypub section ahead
Else it is too easy to mistake for another MRF policy.