Drop activity_type override for uploads
Afaict this was never used, but keeping this (in theory) possible
hinders detecting which objects are actually media uploads and
which proper ActivityPub objects.

It was originally added as part of upload support itself in
02d3dc6869f388388ea744ea4ee3b54279d55e86 without being used
and `git log -S:activity_type` and `git log -Sactivity_type:`
don't find any other commits using this.